aider: Created retriever.py file with code to search the myscale vectorstore for similarity search and provide meta-data as sources from user uploaded documents.
# Context:
USER: wonerful! are you able to cd /Users/matthewsimon/Documents/GitHub/Medex-Public-MITP/Medex-Public-MITP/Medex, and create a new retriever.py file that contains the necessary code to perform the following: a. search the myscale vectorstore for similarity search in order to find the appropriate response to the user query. b. provide meta-data as sources from the user uploaded documents.