aider: Added a new function `query_retriever` to retrieve and print search results in a condensed format.
# Context:
USER: be sure to include the actual query function as well. i need to be able to be continously call the model throughout the chat, so i need to ensure there is a condensed block container the query area that is easily repeatable in the terminal, rather than a giant codeblock. If you want to implement any visual tools to enhance the appearance of the terminal here as well- that would a nice surprise.